requirements.txt
bs4
requests
newspaper3k
nltk
requests_html
html5lib